moved controller and service methods to util classes
