+++ src/main/java/com/example/demo/controller/RegistrationRestController.java
@@ -1,247 +1,34 @@
 package com.example.demo.controller;
 
-import com.example.demo.dto.PasswordDto;
-import com.example.demo.dto.UserDto;
-import com.example.demo.error.InvalidOldPasswordException;
 import com.example.demo.error.UserAlreadyExistException;
-import com.example.demo.persistence.model.User;
-import com.example.demo.persistence.model.VerificationToken;
-import com.example.demo.security.ISecurityUserService;
-import com.example.demo.service.IUserService;
+import com.example.demo.persistence.model.dto.PasswordDto;
+import com.example.demo.persistence.model.dto.UserDto;
 import com.example.demo.util.GenericResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.context.MessageSource;
-import org.springframework.core.env.Environment;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.validation.Valid;
 import java.io.UnsupportedEncodingException;
-import java.time.Instant;
-import java.util.Date;
 import java.util.Locale;
-import java.util.Optional;
-import java.util.UUID;
 
-@RestController
-//@CrossOrigin(origins = "http://localhost:3000")
-public class RegistrationRestController {
+public interface RegistrationRestController {
 
-    private final Logger LOGGER = LoggerFactory.getLogger(getClass());
+    GenericResponse registerUserAccount(UserDto accountDto, HttpServletRequest request) throws UserAlreadyExistException;
 
-    @Autowired
-    private IUserService userService;
+    GenericResponse activateUserAccount(String token, HttpServletRequest request);
 
-    private ISecurityUserService securityUserService;
+    GenericResponse sendRegistrationToken(HttpServletRequest request, final String existingToken);
 
-    @Autowired
-    private MessageSource messages;
+    GenericResponse recoverPassword(HttpServletRequest request, String userEmail);
 
-    @Autowired
-    private JavaMailSender mailSender;
+    GenericResponse resetPassword(Locale locale, PasswordDto passwordDto);
 
-    @Autowired
-    private ApplicationEventPublisher eventPublisher;
+    GenericResponse updatePassword(Locale locale, PasswordDto passwordDto);
 
-    @Autowired
-    private Environment env;
+    GenericResponse modifyUser2FA(boolean use2FA) throws UnsupportedEncodingException;
 
-    public RegistrationRestController() {
-        super();
-    }
+    GenericResponse inactivateAccount(String verificationToken);
 
-    // endpoint to test if jwt token auth is working
-    @GetMapping("/test")
-    public String getTest() {
-        return "Test Works";
-    }
+    GenericResponse deactivateLostAccount(String email);
 
-    // Registration
-    @PostMapping("/user/registration/verification")
-    public GenericResponse registerUserAccount(@RequestBody @Valid final UserDto accountDto, final HttpServletRequest request) throws UserAlreadyExistException {
-        LOGGER.debug("Registering user account with information: {}", accountDto);
-
-        try {
-            final User registered = userService.registerNewUserAccount(accountDto);
-            final String token = UUID.randomUUID().toString();
-            mailSender.send(sendAccountActivationEmail(token, registered.getEmail(), registered));
-            return new GenericResponse("success");
-        } catch ( UserAlreadyExistException userAlreadyExistException ) {
-            return new GenericResponse(messages.getMessage("message.regError", null, request.getLocale()));
-        }
-        // Will work on this later on
-        // userService.addUserLocation(registered, getClientIP(request));
-        // eventPublisher.publishEvent(new OnRegistrationCompleteEvent(registered, request.getLocale(), (String) getAppUrl(request)));
-    }
-
-    @GetMapping("/user/registration/activation")
-    public GenericResponse activateUserAccount(@RequestParam final String token, final HttpServletRequest request) {
-        final VerificationToken verificationToken = userService.getVerificationToken(token);
-        if (verificationToken.getExpiryDate().before(Date.from(Instant.now()))) {
-            return new GenericResponse(messages.getMessage("message.expired", null, request.getLocale()));
-        } else {
-            final User user = userService.getUser(verificationToken.getToken());
-            userService.activateAccount(user);
-            return new GenericResponse("success");
-        }
-    }
-
-    // User activation - verification
-    @GetMapping("/user/resendRegistrationToken")
-    public GenericResponse sendRegistrationToken(final HttpServletRequest request, @RequestParam("token") final String existingToken) {
-        final VerificationToken newToken = userService.generateNewVerificationToken(existingToken);
-        final User user = userService.getUser(newToken.getToken());
-        mailSender.send(constructResendVerificationTokenEmail((getAppUrl(request).toString()), request.getLocale(), newToken, user));
-        return new GenericResponse(messages.getMessage("message.resendToken", null, request.getLocale()));
-    }
-
-    // Send link to email after forgetting password to reset password. The link will redirect user to /user/savePassword
-//    @PostMapping("/user/registration/sendOneTimePasswordLink")
-//    public GenericResponse recoverPassword(final HttpServletRequest request, @RequestParam("email") final String userEmail) {
-//        final User user = userService.findUserByEmail(userEmail);
-//        if (user != null) {
-//            final String token = UUID.randomUUID().toString();
-//            userService.createPasswordResetTokenForUser(user, token);
-//            mailSender.send(constructResetTokenEmail(getAppUrl(request).toString(), request.getLocale(), token, user));
-//        }
-//        return new GenericResponse(messages.getMessage("message.resetPasswordEmail", null, request.getLocale()));
-//    }
-
-
-    @PostMapping("/user/registration/oneTimePassword")
-    public GenericResponse recoverPassword(final HttpServletRequest request, @RequestParam("email") final String userEmail) {
-
-        final User user = userService.findUserByEmail(userEmail);
-
-        if (user != null) {
-            final String tokenString = UUID.randomUUID().toString();
-            userService.createPasswordResetTokenForUser(user, tokenString);
-            final String oneTimePassword = userService.generateOneTimePassword();
-            userService.changeUserPassword(user, oneTimePassword, (60 * 24));
-            SimpleMailMessage mailMessage = constructEmail("Your one time password!", "Your one time password is: " + oneTimePassword + "\n\nThank You!\nChrist Covenant Church\n(314) 839-0292", user);
-            mailSender.send(mailMessage);
-        }
-
-        return new GenericResponse("Your one time password has been sent to your email!");
-    }
-
-    // Reset password after forgetting old password
-    @PostMapping("/user/resetPassword")
-    public GenericResponse resetPassword(final Locale locale, @RequestBody @Valid PasswordDto passwordDto) {
-
-        final String result = securityUserService.validatePasswordResetToken(passwordDto.getToken());
-
-        if (result != null) {
-            return new GenericResponse(messages.getMessage("auth.message." + result, null, locale));
-        }
-
-        Optional<User> user = userService.getUserByPasswordResetToken(passwordDto.getToken());
-        if (user.isPresent()) {
-            userService.changeUserPassword(user.get(), passwordDto.getNewPassword(), (365 * 60 * 24));
-            return new GenericResponse(messages.getMessage("message.resetPasswordSuc", null, locale));
-        } else {
-            return new GenericResponse(messages.getMessage("auth.message.invalid", null, locale));
-        }
-    }
-
-    // Change User password from old password
-    @PostMapping("/user/updatePassword")
-    public GenericResponse updatePassword(@RequestBody final Locale locale, @RequestBody @Valid PasswordDto passwordDto) {
-        final User user = userService.findUserByEmail(((User) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getEmail());
-        if (!userService.checkIfValidOldPassword(user, passwordDto.getOldPassword())) {
-            throw new InvalidOldPasswordException();
-        }
-        userService.changeUserPassword(user, passwordDto.getNewPassword(), (60 * 24));
-        return new GenericResponse(messages.getMessage("message.updatePasswordSuc", null, locale));
-    }
-
-    // Change user 2 factor authentication
-    @PostMapping("/user/update/2fa")
-    public GenericResponse modifyUser2FA(@RequestParam("use2FA") final boolean use2FA) throws UnsupportedEncodingException {
-        final User user = userService.updateUser2FA(use2FA);
-        if (use2FA) {
-            return new GenericResponse(userService.generateQRUrl(user));
-        }
-        return null;
-    }
-
-    // Inactivate account
-    @GetMapping("/user/inactivateAccount")
-    public GenericResponse inactivateAccount(@RequestParam final String verificationToken) {
-        final User user = userService.getUser(verificationToken);
-        if (userService.deactivateAccount(verificationToken))
-            return new GenericResponse("success");
-        else
-            return new GenericResponse("Error!", "account deactivation did not work");
-    }
-
-    // IMS-10 Validate and inactivate user
-    @PostMapping("/user/deactivateAccount")
-    public GenericResponse deactivateLostAccount(@RequestParam final String email) {
-        final User user = userService.findUserByEmail(email);
-        final String token = UUID.randomUUID().toString();
-        mailSender.send(sendAccountDeActivationEmail(token, user.getEmail(), user));
-        return new GenericResponse("success");
-    }
-
-    @GetMapping("user/activateAccount")
-    public GenericResponse activateAccount(@RequestParam final String email) {
-        final User registered = userService.findUserByEmail(email);
-        final String token = UUID.randomUUID().toString();
-        mailSender.send(sendAccountActivationEmail(token, registered.getEmail(), registered));
-        return new GenericResponse("success");
-    }
-
-    private SimpleMailMessage sendAccountActivationEmail(final String token, final String email, final User user) {
-        final String subject = "Activate your account!";
-        final String url = "http://localhost:8081/user/registration/activation?token=";
-        final String body = url + userService.createVerificationTokenForUser(user, token).getToken();
-        return constructEmail(subject, body, user);
-    }
-
-    private SimpleMailMessage sendAccountDeActivationEmail(final String token, final String email, final User user) {
-        final String subject = "To deactivate your account, press the link below!";
-        final String url = "http://localhost:8081/user/inactivateAccount?verificationToken=";
-        final String body = url + userService.createVerificationTokenForUser(user, token).getToken();
-        return constructEmail(subject, body, user);
-    }
-
-    private SimpleMailMessage constructResetTokenEmail(final String contextPath, final Locale locale, final String token, final User user) {
-        final String url = contextPath + "/user/changePassword?token=" + token;
-        final String message = messages.getMessage("message.resetPassword", null, locale);
-        return constructEmail("Reset Password", message + " \r\n" + url, user);
-    }
-
-    private SimpleMailMessage constructResendVerificationTokenEmail(final String contextPath, final Locale locale, final VerificationToken newToken, final User user) {
-        final String confirmationUrl = contextPath + "/user/registration/activation?token=" + newToken.getToken();
-        final String message = messages.getMessage("message.resendToken", null, locale);
-        return constructEmail("Resend Registration Token", message + " \r\n" + confirmationUrl, user);
-    }
-
-    private SimpleMailMessage constructEmail(String subject, String body, User user) {
-        final SimpleMailMessage simpleMailMessage = new SimpleMailMessage();
-        simpleMailMessage.setSubject(subject);
-        simpleMailMessage.setText(body);
-        simpleMailMessage.setTo(user.getEmail());
-        simpleMailMessage.setFrom(env.getProperty("spring.mail.username"));
-        return simpleMailMessage;
-    }
-
-    private Object getAppUrl(HttpServletRequest request) {
-        return "http://" + request.getServerName() + ":" + request.getServerPort() + request.getContextPath();
-    }
-
-    private String getClientIP(HttpServletRequest request) {
-        final String xfHeader = request.getHeader("X-Forwarded-For");
-        if (xfHeader == null) {
-            return request.getRemoteAddr();
-        }
-        return xfHeader.split(",")[0];
-    }
+    GenericResponse activateAccount(String email);
 }
+
