--- src/main/java/com/example/demo/controller/UserController.java
@@ -1,37 +1,0 @@
-//package com.example.demo.controller;
-//
-//import com.example.demo.security.ActiveUserStore;
-//import com.example.demo.service.IUserService;
-//import org.springframework.beans.factory.annotation.Autowired;
-//import org.springframework.ui.Model;
-//import org.springframework.web.bind.annotation.GetMapping;
-//import org.springframework.web.bind.annotation.RestController;
-//
-//import java.util.Locale;
-//
-//@RestController
-//public class UserController {
-//
-//    @Autowired
-//    ActiveUserStore activeUserStore;
-//
-//    @Autowired
-//    IUserService userService;
-//
-//    @GetMapping("/")
-//    public String home() {
-//        return "Welcome!";
-//    }
-//
-//    @GetMapping("/loggedUsers")
-//    public String getLoggedUsers(final Locale locale, final Model model) {
-//        model.addAttribute("users", activeUserStore.getUsers());
-//        return "users";
-//    }
-//
-//    @GetMapping("/loggedUsersFromSessionRegistry")
-//    public String getLoggedUsersFromSessionRegistry(final Locale locale, final Model model) {
-//        model.addAttribute("users", userService.getUsersFromSessionRegistry());
-//        return "users";
-//    }
-//}
