+++ src/main/java/com/example/demo/persistence/model/dto/UserDto.java
@@ -1,4 +1,4 @@
-package com.example.demo.dto;
+package com.example.demo.persistence.model.dto;
 
 import com.example.demo.validation.PasswordMatches;
 import com.example.demo.validation.ValidEmail;
