+++ src/main/resources/application.properties
@@ -2,9 +2,9 @@ server.port=8081
 ### Hibernate Properties
 ## The SQL dialect makes Hibernate generate better SQL for the chosen database
 #spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5InnoDBDialect
-spring.datasource.url=jdbc:mysql://localhost:3306/test?jdbcCompliantTruncation=false
+spring.datasource.url=jdbc:mysql://localhost:3306/ims?jdbcCompliantTruncation=false
 spring.datasource.username=root
-spring.datasource.password=PassSomeCash$1@
+spring.datasource.password=root
 ## Hibernate Properties
 # The SQL dialect makes Hibernate generate better SQL for the chosen database
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5InnoDBDialect
